Finished Series

+++ app/Http/Controllers/API/UserController.php
@@ -21,7 +21,10 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        if(\Gate::allows('isAdmin')){
+            return User::latest()->paginate(5);
+        }
+
     }
 
     /**
@@ -123,8 +126,21 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['message' => 'user deleted'];
+    }
+
+    public function search(){
+        if($search = \Request::get('q')){
+            $users = User::where(function($query) use ($search){
+                $query->where('name','LIKE',"%$search%")
+                        ->orWhere('email','LIKE', "%$search%");
+            })->paginate(5);
+        }else{
+            $users = User::laters()->paginate(5);
+        }
+        return $users;
     }
 }
